Relative image URL's are not resolved correctly!
My JSON image list contains *relative* image URLs.  These do not display in the imagebrowser window, but if I manage to click on the tiny 2x2 pixel dot standing in for the image, it gets successfully added to CKeditor's image dialog.

The problem is that browser.js needs to know the value of the CKeditor's config.baseHref.  This modification sends it in the querystring.