If the CKeditor instance has config.baseHref set, set <base href="..."> accordingly on the imagebrowser document.
Before doing this, it's necessary to protect the stylesheet URL from being affected.  This is done by setting the link href to an absolute URL (gotten by replacing everything after the last "/" in location.href with "browser.css").

I also made the "thumb" propery optional in the server's JSON objects, in case there are no thumbnails for the images.  If missing or empty, it just defaults to the value of "image".