Minor CSS fix (unrelated to the <base> tag issue).
If you want to see those rounded corners, you need to ensure the image is clipped properly - hence "overflow:hidden".  Otherwise the corners of the image cover the rounded parts of the border.

I also ensured that the thumbnails are scaled down if either dimension is greater than 200px, while still maintaining aspect ratio.  This is particularly useful if you don't use thumbnails at all, just full-size images.

Note that mslinn's CSS update also does this, but not the rounded corners fix.  (Plus he does a lot of other stuff also).

So actually, if implementing mslinn's stylesheet, just add "overflow:hidden" to the .thumbnail class and forget about this update.

Otherwise, this is kind of a minimum-change CSS fix.